+++ src/components/CourseInfo/course-info.css
@@ -36,8 +36,15 @@
 }
 
 .course-description {
-    width: 60%;
+    width: 55vw;
     margin-left: 30px;
+    margin-right: 20px;
+    text-align: justify;
+    text-justify: inter-word;
+}
+
+.course-description p {
+    margin-right: 20px;
 }
 
 .course-details {
