+++ src/store/authors/authorsThunk.js
@@ -1,0 +1,14 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import { getAllAuthorsService } from '../../services';
+
+export const fetchAuthorsData = createAsyncThunk(
+	'authors/fetchAuthorsData',
+	async () => {
+		try {
+			const data = await getAllAuthorsService();
+			return data;
+		} catch (error) {
+			throw error;
+		}
+	}
+);
