+++ src/store/courses/coursesThunk.js
@@ -1,0 +1,14 @@
+import { createAsyncThunk } from '@reduxjs/toolkit';
+import { getAllCoursesService } from '../../services';
+
+export const fetchCoursesData = createAsyncThunk(
+	'courses/fetchCoursesData',
+	async () => {
+		try {
+			const data = await getAllCoursesService();
+			return data;
+		} catch (error) {
+			throw error;
+		}
+	}
+);
